movie_admin/requirements-base.txt
django==3.1
django-model-utils==4.0.0 # Набор полезных базовых классов и утилит для Django
psycopg2-binary==2.8.6
wheel==0.36.2
django_extensions==3.1.0